requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
django-ckeditor==6.7.0
django-js-asset==2.1.0
idna==3.4
pyTelegramBotAPI==4.12.0
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.4
django-shortcuts